requirements.txt
asgiref==3.3.1
Django==3.1.7
django-phone-field==1.8.1
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
